fix: ConfigurationActivity Bug Fix & Delete Widget
In ACTIVE Configuration Mode, the system considers all the pending
intents to be identical since they all have the same target class i.e.
ConfigurationActivity. Thus, the pending intents are updated for all
the widget instances while updating the pending intent of anyone of
them, irrespective of the fact that they have different intent extras.

To solve this, a dummy action is added for each widget. This action is
identical for a particular widget instance but different for different
widget instances.

Also, a workaround code to delete widget since onDeleted is not called
while deleting a widget.