feat: Show Active Widgets Info in MainActivity
Also, newly added widget will auto close after setting the package name
to be this.app and configuration mode will be turned on. This means that
on adding a new widget all old widgets too will become visible and you
will have to close the configuration mode to hide the widgets.