feat: Cache all apps in database
The app now stores app names and package names in an SQLite database.
The database is deleted and created again every time an app is installed
or uninstalled on the device. This database operation happens on the
background thread. The scrolling of the list of apps is not smooth as
the app icons are not cached. This will be fixed in the next commit.