fix: Hotfix for v-1.3
This version was done to make quick bug fixes as reported by users.
When a package(app) is removed from the system, the widget now becomes a
blank widget as compared to the erratic behavior previously.

Also a design decision was taken to always launch configuration activity
in active config mode. If it is launched from home screen, it is
definitely in active config mode but if it's launched from clicking a
list item from the main activity, the config mode is turned on before
launching the config activity.